Restructured routes.js into Controllers and Routes.
