+++ package.json
@@ -11,13 +11,13 @@
     "cookie-parser": "^1.4.3",
     "dotenv": "^4.0.0",
     "ejs": "^2.3.2",
-    "express": "^4.15.5",
+    "express": "^4.16.3",
     "express-mailer": "^0.3.1",
     "jsonwebtoken": "^8.2.0",
     "morgan": "^1.7.0",
     "mysql": "^2.15.0",
     "mysql2": "^1.5.2",
     "randomstring": "^1.1.5",
-    "sequelize": "^4.35.2"
+    "sequelize": "^4.37.1"
   }
 }
